requirements.txt
langchain==0.2.10
langchain-community==0.2.9
langchain-openai==0.1.7
sentence-transformers==2.7.0
faiss-cpu==1.8.0
streamlit==1.38.0
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken==0.7.0
openai==1.40.0
